break out errorf error formatter into new file
